Fix: Final Lint Error Corrections and Hook Refactoring
This commit resolves the final set of linting errors you reported, primarily focusing on fixing 'Fast Refresh' warnings and cleaning up imports.

The following structural changes were made to address the errors:
- I created the `src/hooks` directory to better separate concerns and resolve Fast Refresh warnings.
- I moved the `useAuth` hook from `AuthContext.jsx` to `src/hooks/useAuth.js`.
- I moved the `useToast` hook from `ToastContext.jsx` to `src/hooks/useToast.js`.
- I refactored `AuthContext.jsx` and `ToastContext.jsx` to only have the Provider as the default export.
- I updated all components using these hooks to import from the new location.
- I removed the unused `useEffect` import from `ToastContext.jsx`.

These changes result in a cleaner, better structured, and lint-free codebase.
