+++ src/App.jsx
@@ -1,5 +1,5 @@
 import { Routes, Route, Navigate } from 'react-router-dom';
-import { useAuth } from './contexts/AuthContext';
+import { useAuth } from './hooks/useAuth';
 
 import Layout from './components/Layout';
 import DashboardPage from './pages/DashboardPage';
