+++ src/hooks/useAuth.js
@@ -1,0 +1,6 @@
+import { useContext } from 'react';
+import { AuthContext } from '../contexts/AuthContext';
+
+export const useAuth = () => {
+  return useContext(AuthContext);
+};
