+++ src/hooks/useToast.js
@@ -1,0 +1,6 @@
+import { useContext } from 'react';
+import { ToastContext } from '../contexts/ToastContext';
+
+export const useToast = () => {
+  return useContext(ToastContext);
+};
